Make the runtime artifacts MPP
