+++ gradle.properties
@@ -1,13 +1,15 @@
 kotlin.code.style=official
 kotlin.jvm.addClassesVariant=true
+kotlin.js.yarn=false
 org.gradle.parallel=true
-org.gradle.jvmargs=-Xmx2048m
 org.gradle.caching=true
 org.gradle.configuration-cache=true
 org.gradle.configuration-cache.parallel=true
 org.gradle.configuration-cache.entries-per-key=100
-org.gradle.configureondemand=true
-org.gradle.unsafe.isolated-projects=true
+org.gradle.unsafe.isolated-projects=false
+kotlin.kmp.isolated-projects.support=enable
+kotlin.native.ignoreDisabledTargets=true
+org.gradle.jvmargs=-XX:MaxMetaspaceSize=1g -Xmx4096m
 org.gradle.kotlin.dsl.skipMetadataVersionCheck=false
 org.gradle.kotlin.dsl.dcl=true
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
@@ -15,3 +17,4 @@ org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
 org.jetbrains.dokka.experimental.tryK2=true
 org.jetbrains.dokka.experimental.tryK2.nowarn=true
 group=io.github.hfhbd.kfx
+kotlin.native.enableKlibsCrossCompilation=true
