+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-kotlin = "2.1.21"
+kotlin = "2.2.0-RC"
 ktor = "3.1.3"
 serialization = "1.8.1"
