+++ ktor-server-runtime/build.gradle.kts
@@ -1,10 +1,12 @@
 plugins {
-    id("jvmModule")
+    id("runtimeModule")
 }
 
-dependencies {
-    api(libs.serialization.core)
-    api(libs.ktor.server.core)
+kotlin.sourceSets.commonMain {
+    dependencies {
+        api(libs.serialization.core)
+        api(libs.ktor.server.core)
+    }
 }
 
 licensee {
