+++ oauth2-runtime/src/commonMain/kotlin/io/github/hfhbd/kfx/oauth2/OAuth2Token.kt
@@ -1,0 +1,18 @@
+package io.github.hfhbd.kfx.oauth2
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+public data class OAuth2Token(
+    @SerialName("access_token")
+    val accessToken: String,
+    @SerialName("token_type")
+    val tokenType: String,
+    @SerialName("expires_in")
+    val expiresIn: Long? = null,
+    @SerialName("refresh_token")
+    val refreshToken: String? = null,
+    val scope: String? = null,
+    val expired: Boolean = false,
+)
