+++ soap-runtime/build.gradle.kts
@@ -1,9 +1,11 @@
 plugins {
-    id("jvmModule")
+    id("runtimeModule")
 }
 
-dependencies {
-    api(libs.serialization.core)
-    api(libs.serialization.xml)
-    api(libs.datetime)
+kotlin.sourceSets.commonMain {
+    dependencies {
+        api(libs.serialization.core)
+        api(libs.serialization.xml)
+        api(libs.datetime)
+    }
 }
